+++ src/main/java/com/training/java/transaction/tracker/service/response/CreateTransactionResponse.java
@@ -2,20 +2,20 @@ package com.training.java.transaction.tracker.service.response;
 
 public class CreateTransactionResponse extends BaseResponse {
 
-    private final int transactionId;
+    private final long transactionId;
 
     public CreateTransactionResponse(boolean successful, String description) {
         super(successful, description);
         transactionId = -1;
     }
 
-    public CreateTransactionResponse(boolean successful, String description, int transactionId) {
+    public CreateTransactionResponse(boolean successful, String description, long transactionId) {
         super(successful, description);
         this.transactionId = transactionId;
     }
 
 
-    public int getIdentifier() {
+    public long getIdentifier() {
         return transactionId;
     }
 }
